Abstracted snippet loading
Adds a utility script that is used to load the snippets (more coming).
